pyproject.toml
[tool.poetry]
name = "postgres_explain_analyze_json_schema"
version = "0.1.0"
description = ""
authors = ["Steven Kalt <kalt.steven@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
genson = "^1.2.2"
PyYAML = "^6.0"
jsonschema = "^4.16.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
